pyproject.toml
[tool.poetry]
name = "derive-client"
version = "0.2.10"
description = ""
authors = ["8baller <8baller@station.codes>"]
readme = "README.md"
packages = [{include = "derive"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
requests = "^2"
web3 = "^6"
websocket-client = ">=0.32.0,<1"
setuptools = ">=68.2.2,<80"
rich-click = "^1.7.1"
python-dotenv = ">=0.14.0,<0.18.0"
pandas = ">=1,<=3"


[tool.poetry.scripts]
derive= "derive.cli:cli"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
tbump = "^6.11.0"
pytest-rerunfailures = "^13.0"
semver = ">=2.9.1,<3.0.0"
numpy = "<2"

mkdocs = "^1.3.1"
mkdocs-include-markdown-plugin = "^3.6.1"
mkdocs-material = "^8.4.0"
mkdocs-material-extensions = "^1.0.3"
mkdocs-autorefs = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py37', 'py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
